EVAA/requirements.txt
# EVAA - Electra Voice AI Agent Requirements

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
motor==3.3.2
beanie==1.23.6
pymongo==4.6.0

# AI and ML
pydantic-ai==0.0.14
groq==0.4.1

# Speech and Audio
deepgram-sdk==3.2.7
pydub==0.25.1

# Configuration
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
